requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
dj-database-url
python-decouple
google-generativeai
whitenoise
gunicorn
